// src/server/models/Product.ts
export interface Product {
  id?: string;
  code: string;
  name: string;
  packing?: string;
  boxPack?: string;
  casePack?: string;
  printMark?: string;
  hsn?: string;
  company?: string;
  rackNo?: string;
  mark?: string;
  sch?: string;
  aiodCode?: string;
  marketedBy?: string;
  category?: string;
  saleQtyMin?: string;
  max?: string;
  orderLevel?: string;
  orderQty?: string;
  boxQtyRates?: string;
  boxPurPrice?: string;
  comment?: string;
  discontinued?: boolean;
  purchaseQty?: string;
  saleQty?: string;
  purchaseFree?: string;
  saleFree?: string;
  drugFormula?: string;
  createdAt?: Date;
  updatedAt?: Date;
}